Prevents Bad State: No Element from being thrown
If the reduce method is called on an empty stream (like Stream.empty()), a Bad State: No Element exception is thrown. This change adds a check to determine if the stream is empty before sending the request. If the stream is empty, the request is sent but if the stream is not empty, the stream is reduced before being sent.

I ran into this issue when using doing a get request on flutter web.