⚡️ Mark `DioMixin`'s todo and add tests (#2066)
`DioMixin` was never a real mixin before. Marking it to use `mixin
class` once we use `>=3.0.0` as the SDK constraint.

### New Pull Request Checklist

- [x] I have read the
[Documentation](https://pub.dev/documentation/dio/latest/)
- [x] I have searched for a similar pull request in the
[project](https://github.com/cfug/dio/pulls) and found none
- [x] I have updated this branch with the latest `main` branch to avoid
conflicts (via merge from master or rebase)
- [x] I have added the required tests to prove the fix/feature I'm
adding
- [x] I have updated the documentation (if necessary)
- [x] I have run the tests without failures
- [ ] I have updated the `CHANGELOG.md` in the corresponding package

### Additional context and info (if any)

The following exception will be raised if we roll the lower bound of the
package to ^3.0.0:
`The class 'DioMixin' can't be used as a mixin because it's neither a
mixin class nor a mixin.`