Converge melos usage and CI scripts & add code coverage (#2136)
The idea of this PR is to make it easier to run tests locally based on
melos while also adding code coverage reporting.

* move all scripts to melos and make them available for usage locally
* automatically generation coverage information for all Dart tests (vm,
chrome, firefox)
* print coverage results to CLI for stable SDK
* only run analyze/format/publish on Stable SDK, this way we can skip
the patching for format/analyze for older SDK versions which actually
makes no sense since it is not being released


I want to further extend the coverage reporting by using
https://github.com/clearlyip/code-coverage-report-action.
This action would run for each push/merge to the main branch an generate
a constant coverage baseline for the main branch. The baseline is
uploaded as workflow artifact and as such available to pull request
which can then display the change in coverage.

<!-- Write down your pull request descriptions. -->

### New Pull Request Checklist

- [ ] I have read the
[Documentation](https://pub.dev/documentation/dio/latest/)
- [ ] I have searched for a similar pull request in the
[project](https://github.com/cfug/dio/pulls) and found none
- [ ] I have updated this branch with the latest `main` branch to avoid
conflicts (via merge from master or rebase)
- [ ] I have added the required tests to prove the fix/feature I'm
adding
- [ ] I have updated the documentation (if necessary)
- [ ] I have run the tests without failures
- [ ] I have updated the `CHANGELOG.md` in the corresponding package

### Additional context and info (if any)

<!-- Provide more context and info about the PR. -->