Fix coverage comment from forks
When the coverage_comment workflow is triggered from a PR that originates from a fork, then the workflow event does not contain the PR number. But we need the PR number to attach the sticky coverage comment. So workaround this by storing the PR number in the workflow artifact that also contains the coverage result.