add Utf8JsonTransformer (#2239)
<!-- Write down your pull request descriptions. -->

### New Pull Request Checklist

- [x] I have read the
[Documentation](https://pub.dev/documentation/dio/latest/)
- [x] I have searched for a similar pull request in the
[project](https://github.com/cfug/dio/pulls) and found none
- [x] I have updated this branch with the latest `main` branch to avoid
conflicts (via merge from master or rebase)
- [x] I have added the required tests to prove the fix/feature I'm
adding
- [x] I have updated the documentation (if necessary)
- [x] I have run the tests without failures
- [x] I have updated the `CHANGELOG.md` in the corresponding package

### Additional context and info (if any)

fixes #2238 

<!-- Provide more context and info about the PR. -->

This PR adds the `Utf8JsonTransformer`.

I made sure that it is as close to `BackgroundTransformer` as possible.
The only behavior change is that is checks to Content-Length header for
deciding whether to switch to a background isolate for decoding or not.
If the Content-Lenght header is not set or not valid, it will fall back
to counting the bytes in the response.

This is because it's slightly faster to feed to the response stream
directly into the decoder without consolidating all the response bytes
into a single Uint8List first. In case the server does not send a valid
Content-Length header, the Transformer will check the actual bytes from
the response, as the `BackgroundTransformer` does.


The remaining question is if we should set the Transformer as default. 
IMO it would make sense to change the default transformer to

```dart
  /// The default [Transformer] that transfers requests and responses
  /// into corresponding content to send.
  @override
  Transformer transformer = Utf8JsonTransformer(contentLengthIsolateThreshold: 50 * 1024);
```

so all users get the improved performance by default.

---------

Signed-off-by: Jonas Uekötter <ueman@users.noreply.github.com>
Co-authored-by: Jonas Uekötter <ueman@users.noreply.github.com>