Fix: Clarify the behavior of concurrent requests in QueuedInterceptor (#2347)
The previous documentation for `QueuedInterceptor` could lead to misunderstandings about how concurrent requests are handled. The explanation suggested that requests would have to wait for the previous request to complete before being initiated.

This update clarifies that requests are processed in three independent queues: `onRequest`, `onResponse`, and `onError`. These queues run in parallel.

---------

Signed-off-by: 友人A <58846244+Passer-by@users.noreply.github.com>
Co-authored-by: Alex Li <github@alexv525.com>